Move classes in helpers.py to objects.py.
