+++ farcy/helpers.py
@@ -1,16 +1,8 @@
 """Helper methods and classes."""
 
-
 from collections import defaultdict
-try:
-    from configparser import ConfigParser  # PY3
-    basestring = str
-except ImportError:
-    from ConfigParser import SafeConfigParser as ConfigParser  # PY2
-from datetime import timedelta, tzinfo
 from github3 import GitHub
 from github3.exceptions import GitHubError
-import logging
 import os
 import sys
 from .const import FARCY_COMMENT_START, NUMBER_RE, CONFIG_DIR
@@ -190,121 +182,3 @@ def subtract_issues_by_line(by_line, by_line2):
         if filtered:
             result[key] = filtered
     return result
-
-
-class Config(object):
-
-    """Holds configuration for Farcy."""
-
-    ATTRIBUTES = {'debug', 'exclude_paths', 'limit_users', 'log_level',
-                  'pr_issue_report_limit', 'pull_requests', 'start_event'}
-    LOG_LEVELS = {'CRITICAL', 'ERROR', 'WARNING', 'INFO', 'DEBUG', 'NOTSET'}
-    PATH = os.path.join(CONFIG_DIR, 'farcy.conf')
-
-    @property
-    def log_level_int(self):
-        """Int value of the log level."""
-        return getattr(logging, self.log_level)
-
-    @property
-    def session(self):
-        """Return GitHub session. Create if necessary."""
-        if self._session is None:
-            self._session = get_session()
-        return self._session
-
-    def __init__(self, repository, **overrides):
-        """Initialize a config with default values."""
-        self._session = None
-        self.repository = repository
-        self.set_defaults()
-        self.load_config_file()
-        self.override(**overrides)
-
-    def __repr__(self):
-        """String representation of the config."""
-        keys = sorted(x for x in self.__dict__ if not x.startswith('_')
-                      and x != 'repository')
-        arg_fmt = ', '.join(['{0}={1!r}'.format(key, getattr(self, key))
-                             for key in keys])
-        return 'Config({0!r}, {1})'.format(self.repository, arg_fmt)
-
-    def __setattr__(self, attr, value):
-        """
-        Set new config attribute.
-
-        Validates new attribute values and tracks if changed from default.
-
-        """
-        if attr == 'debug' and parse_bool(value):
-            # Force log level when in debug mode
-            setattr(self, 'log_level', 'DEBUG')
-        elif attr in ('exclude_paths', 'pull_requests'):
-            if value is not None:
-                value = parse_set(value)
-        elif attr == 'limit_users':
-            if value:
-                value = parse_set(value, normalize=True)
-        elif attr == 'log_level' and self.debug:
-            return  # Don't change level in debug mode
-        elif attr == 'log_level' and value is not None:
-            value = value.upper()
-            if value not in self.LOG_LEVELS:
-                raise FarcyException('Invalid log level: {0}'.format(value))
-        elif attr == 'repository' and value is not None:
-            repo_parts = value.split('/')
-            if len(repo_parts) != 2:
-                raise FarcyException('Invalid repository: {0}'.format(value))
-        elif attr in ('pr_issue_report_limit', 'start_event'):
-            if value is not None:
-                value = int(value)
-        super(Config, self).__setattr__(attr, value)
-
-    def load_config_file(self):
-        """Load value overrides from configuration file."""
-        if not os.path.isfile(self.PATH):
-            return
-
-        config_file = ConfigParser()
-        config_file.read(self.PATH)
-
-        if not self.repository and \
-                config_file.has_option('DEFAULT', 'repository'):
-            self.repository = config_file.get('DEFAULT', 'repository')
-
-        self.override(**dict(config_file.items(
-            self.repository if config_file.has_section(self.repository)
-            else 'DEFAULT')))
-
-    def override(self, **overrides):
-        """Override the config values passed as keyword arguments."""
-        for attr, value in overrides.items():
-            if attr in self.ATTRIBUTES and value:
-                setattr(self, attr, value)
-
-    def set_defaults(self):
-        """Set the default config values."""
-        self.debug = False
-        self.exclude_paths = None
-        self.limit_users = None
-        self.log_level = 'ERROR'
-        self.pr_issue_report_limit = 128
-        self.pull_requests = None
-        self.start_event = None
-
-    def user_whitelisted(self, user):
-        """Return if user is whitelisted."""
-        return self.limit_users is None or user.lower() in self.limit_users
-
-
-class UTC(tzinfo):
-
-    """Provides a simple UTC timezone class.
-
-    Source: http://docs.python.org/release/2.4.2/lib/datetime-tzinfo.html
-
-    """
-
-    dst = lambda x, y: timedelta(0)
-    tzname = lambda x, y: 'UTC'
-    utcoffset = lambda x, y: timedelta(0)
